Dev/professional/bugfix merge feature

See merge request chengdu/dde-file-manager!57

+++ dde-file-manager-lib/views/dfmopticalmediawidget.cpp
@@ -137,7 +137,7 @@ DFMOpticalMediaWidget::DFMOpticalMediaWidget(QWidget *parent) :
     connect(m_pStatisticWorker, &DFileStatisticsJob::finished, this, [=] {
         DeviceProperty dp = ISOMaster->getDevicePropertyCached(d->getCurrentDevice());
 
-        if (m_pStatisticWorker->totalSize() > dp.avail) {
+        if (static_cast<quint64>(m_pStatisticWorker->totalSize()) > dp.avail) {
             //fix: 光盘容量小于刻录项目，对话框提示：目标磁盘剩余空间不足，无法进行刻录！
             //qDebug() << d->m_selectBurnFilesSize / 1024 / 1024 << "MB" << dp.avail / 1024 / 1024 << "MB";
             DDialog dialog(this);
@@ -149,7 +149,7 @@ DFMOpticalMediaWidget::DFMOpticalMediaWidget(QWidget *parent) :
         }
 
         QScopedPointer<BurnOptDialog> bd(new BurnOptDialog(d->getCurrentDevice(), this));
-        bd->setJobWindowId(this->window()->winId());
+        bd->setJobWindowId(static_cast<int>(this->window()->winId()));
         bd->exec();
 //        if (bd->exec() == DDialog::Accepted) {
 //            // 发送信号停止扫描光驱的计时器
